ml_model.py
import os
import random
import numpy as np
import json
from PIL import Image
import logging

logger = logging.getLogger(__name__)

# This is a placeholder for the actual ML model that would analyze water samples
# In a real implementation, this would use the user's trained model

def analyze_water_sample(image_path):
    """
    Analyze a water sample image to detect copper concentration.
    
    In a real implementation, this would:
    1. Load the image
    2. Preprocess it for the ML model
    3. Run the ML model to predict copper concentration
    4. Classify the result based on risk levels
    
    For this demo, we'll return simulated results.
    
    Args:
        image_path (str): Path to the uploaded image file
        
    Returns:
        dict: Analysis results including concentration, risk level, and details
    """
    logger.debug(f"Analyzing water sample: {image_path}")
    
    try:
        # Load and process the image
        img = Image.open(image_path)
        
        # Extract basic image properties for demonstration
        width, height = img.size
        
        # In a real implementation, you would:
        # 1. Process the image (resize, normalize, etc.)
        # 2. Extract features related to color/fluorescence
        # 3. Run these features through your ML model
        
        # Simulated analysis - in a real app, replace with actual ML model
        # Generate a simulated copper concentration (mg/L)
        concentration = random.uniform(0.1, 3.0)
        concentration = round(concentration, 2)
        
        # Classify the concentration
        risk_level = classify_copper_concentration(concentration)
        
        # Generate detailed analysis data
        details = {
            "image_dimensions": f"{width}x{height}",
            "concentration_unit": "mg/L",
            "prediction_confidence": round(random.uniform(0.75, 0.98), 2),
            "color_analysis": {
                "r_mean": random.randint(100, 200),
                "g_mean": random.randint(100, 200),
                "b_mean": random.randint(100, 200)
            },
            "safe_threshold": 1.0,
            "warning_threshold": 1.3,
            "danger_threshold": 2.0
        }
        
        result = {
            "concentration": concentration,
            "risk_level": risk_level,
            "details": json.dumps(details)
        }
        
        logger.debug(f"Analysis results: {result}")
        return result
        
    except Exception as e:
        logger.error(f"Error analyzing image: {str(e)}")
        raise Exception(f"Failed to analyze image: {str(e)}")

def classify_copper_concentration(concentration):
    """
    Classify copper concentration into risk levels.
    
    In a real implementation, this would use the thresholds from the trained model.
    
    Args:
        concentration (float): Copper concentration in mg/L
        
    Returns:
        str: Risk level classification
    """
    if concentration < 1.0:
        return "Safe"
    elif concentration < 1.3:
        return "Normal"
    elif concentration < 2.0:
        return "Elevated"
    elif concentration < 2.5:
        return "Risky"
    else:
        return "Hazardous"
